Refactor monkey patch "PatchNamedfileNamedDataConverter".
